mejoras hpa, logs, crud

+++ src/entity/product.go
@@ -1,10 +1,10 @@
 package entity
 
 type Product struct {
-	Id          int     `json:"id"`
-	Nombre      string  `json:"nombre"`
-	Descripcion string  `json:"descripcion"`
-	Precio      float64 `json:"precio"`
-	Stock       int     `json:"stock"`
-	IdSku       string  `json:"id_sku"`
+	Id          int    `json:"id"`
+	Nombre      string `json:"nombre"`
+	Descripcion string `json:"descripcion"`
+	Precio      int64  `json:"precio"`
+	Stock       int64  `json:"stock"`
+	IdSku       string `json:"id_sku"`
 }
